Add dump_lastchar function
The dump_lastchar function was created to help me figure out a bug in my
bookmarks.html. There were stray characters at the end of my URLs,
characters like periods, or right-parenthesis. This function emits the
last character of every URL that the module finds, allowing me to study it.